chore: move node package to sub package of domain
