+++ domain/node/input_node.go
